Add files via upload
This is a website of an online grocery shopping platform developed using HTML, CSS and JS. This project involves only frontend coding languages.  Any feedbacks are welcome on the project. One last thing, we are uploading the images used in this project but we have added the images using …

Topic:javaSript